+++ lib/td_lm/audit.ex
@@ -20,6 +20,21 @@ defmodule TdLm.Audit do
   end
 
   @doc """
+  Publishes `:relations_deprecated` events. Should be called using `Ecto.Multi.run/5`.
+  """
+  def relations_deprecated(_repo, %{deprecated: {_, [_ | _] = relations}}) do
+    relations
+    |> Enum.map(&do_relation_deprecated/1)
+    |> Enum.group_by(&elem(&1, 0), &elem(&1, 1))
+    |> case do
+      %{error: errors} -> {:error, errors}
+      %{ok: event_ids} -> {:ok, event_ids}
+    end
+  end
+
+  def relations_deprecated(_repo, _), do: {:ok, []}
+
+  @doc """
   Publishes a `:tag_deleted` event. Should be called using `Ecto.Multi.run/5`.
   """
   def tag_deleted(_repo, %{tag: tag}, user_id) do
@@ -36,6 +51,11 @@ defmodule TdLm.Audit do
   defp do_relation_deleted(%{source_type: source_type, source_id: source_id} = relation, user_id) do
     payload = Map.take(relation, [:id, :target_id, :target_type, :context])
     publish("relation_deleted", source_type, source_id, user_id, payload)
+  end
+
+  defp do_relation_deprecated(%{source_type: source_type, source_id: source_id} = relation) do
+    payload = Map.take(relation, [:id, :target_id, :target_type, :context])
+    publish("relation_deprecated", source_type, source_id, 0, payload)
   end
 
   defp do_relation_created(
@@ -58,7 +78,7 @@ defmodule TdLm.Audit do
     publish("relation_created", source_type, source_id, user_id, changes)
   end
 
-  defp do_tag_created(%{id: id, value: value} = tag, user_id) do
+  defp do_tag_created(%{id: id, value: _value} = tag, user_id) do
     payload = Map.take(tag, [:value])
     publish("tag_created", "tag", id, user_id, payload)
   end